requirements.txt
pillow
pyttsx3
instabot
